Merge remote-tracking branch 'origin/Tyler's-Version' into Tyler's-Version
Not really sure what github is doing so Jordan's commit may have
disappeared. I saw it on the website but not desktop client. But I made
sure to keep the changes (didn't gut anything this time) to try and fix
the issue. Accounts are now part of the AspNetUsers table. The Manage
page displays user role to confirm that the data is getting passed
around. Next step is to make page viewing restrictions based on user
role.